Started integration of C++ code.

+++ src/components/oscillator.hpp
@@ -46,7 +46,7 @@ namespace live::tritone::vie::processor::component
 
 		uint_fast16_t get_slot_id(const std::string& slot_name) override;
 
-		void set_input_values(uint_fast16_t slot_id, void* values, uint_fast16_t nb_values) override;
+		void set_input_values(uint_fast16_t slot_id, void* values, uint_fast32_t nb_values) override;
 
 		uint_fast32_t get_max_nb_input_values(uint_fast16_t slot_id) override;
